convert event and attribute_change models to mixins
