+++ lib/cia.rb
@@ -1,12 +1,10 @@
 require 'active_record'
 require 'cia/version'
 require 'cia/auditable'
+require 'cia/event_methods'
+require 'cia/attribute_change_methods'
 
 module CIA
-  autoload 'Event', 'cia/event'
-  autoload 'AttributeChange', 'cia/attribute_change'
-  autoload 'SourceValidation', 'cia/source_validation'
-
   class << self
     attr_accessor :exception_handler
     attr_accessor :non_recordable_attributes
